chore: #34 Checkstyle 제외 경로 및 Builder 기본값 추가, 행정구역 패키지명 변경
