+++ src/main/java/com/wholeseeds/mindle/domain/complaint/entity/Complaint.java
@@ -1,7 +1,7 @@
 package com.wholeseeds.mindle.domain.complaint.entity;
 
 import com.wholeseeds.mindle.common.entity.BaseEntity;
-import com.wholeseeds.mindle.domain.location.entity.Subdistrict;
+import com.wholeseeds.mindle.domain.region.entity.Subdistrict;
 import com.wholeseeds.mindle.domain.member.entity.Member;
 import com.wholeseeds.mindle.domain.place.entity.Place;
