--- src/main/java/com/wholeseeds/mindle/domain/location/repository/custom/DistrictRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.location.repository.custom;
-
-public interface DistrictRepositoryCustom {
-}
