--- src/main/java/com/wholeseeds/mindle/domain/location/repository/custom/SubdistrictRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.location.repository.custom;
-
-public interface SubdistrictRepositoryCustom {
-}
