+++ src/main/java/com/wholeseeds/mindle/domain/region/entity/type/SubdistrictType.java
@@ -1,13 +1,14 @@
-package com.wholeseeds.mindle.domain.location.entity.type;
+package com.wholeseeds.mindle.domain.region.entity.type;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CityType {
-	SI("시"),
-	GOON("군");
+public enum SubdistrictType {
+	EUP("읍"),
+	MYEON("면"),
+	DONG("동");
 
 	private final String displayName;
 }
