+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/CityRepository.java
@@ -1,9 +1,9 @@
-package com.wholeseeds.mindle.domain.location.repository;
+package com.wholeseeds.mindle.domain.region.repository;
 
 import java.util.Optional;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepository;
-import com.wholeseeds.mindle.domain.location.entity.City;
+import com.wholeseeds.mindle.domain.region.entity.City;
 
 public interface CityRepository extends JpaBaseRepository<City, Long> {
 	Optional<City> findByName(String name);
