+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/DistrictRepository.java
@@ -1,9 +1,9 @@
-package com.wholeseeds.mindle.domain.location.repository;
+package com.wholeseeds.mindle.domain.region.repository;
 
 import java.util.Optional;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepository;
-import com.wholeseeds.mindle.domain.location.entity.District;
+import com.wholeseeds.mindle.domain.region.entity.District;
 
 public interface DistrictRepository extends JpaBaseRepository<District, Long> {
 	Optional<District> findByName(String name);
