+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/SubdistrictRepository.java
@@ -1,10 +1,10 @@
-package com.wholeseeds.mindle.domain.location.repository;
+package com.wholeseeds.mindle.domain.region.repository;
 
 import java.util.Optional;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepository;
-import com.wholeseeds.mindle.domain.location.entity.City;
-import com.wholeseeds.mindle.domain.location.entity.Subdistrict;
+import com.wholeseeds.mindle.domain.region.entity.City;
+import com.wholeseeds.mindle.domain.region.entity.Subdistrict;
 
 public interface SubdistrictRepository extends JpaBaseRepository<Subdistrict, Long> {
 	Optional<Subdistrict> findByNameAndCity(String name, City city);
