+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/custom/SubdistrictRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.wholeseeds.mindle.domain.region.repository.custom;
+
+public interface SubdistrictRepositoryCustom {
+}
