+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/impl/CityRepositoryImpl.java
@@ -1,9 +1,9 @@
-package com.wholeseeds.mindle.domain.location.repository.impl;
+package com.wholeseeds.mindle.domain.region.repository.impl;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepositoryImpl;
-import com.wholeseeds.mindle.domain.location.entity.City;
-import com.wholeseeds.mindle.domain.location.entity.QCity;
-import com.wholeseeds.mindle.domain.location.repository.custom.CityRepositoryCustom;
+import com.wholeseeds.mindle.domain.region.entity.City;
+import com.wholeseeds.mindle.domain.region.entity.QCity;
+import com.wholeseeds.mindle.domain.region.repository.custom.CityRepositoryCustom;
 
 import jakarta.persistence.EntityManager;
