+++ src/main/java/com/wholeseeds/mindle/domain/region/repository/impl/SubdistrictRepositoryImpl.java
@@ -1,9 +1,9 @@
-package com.wholeseeds.mindle.domain.location.repository.impl;
+package com.wholeseeds.mindle.domain.region.repository.impl;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepositoryImpl;
-import com.wholeseeds.mindle.domain.location.entity.QSubdistrict;
-import com.wholeseeds.mindle.domain.location.entity.Subdistrict;
-import com.wholeseeds.mindle.domain.location.repository.custom.SubdistrictRepositoryCustom;
+import com.wholeseeds.mindle.domain.region.entity.QSubdistrict;
+import com.wholeseeds.mindle.domain.region.entity.Subdistrict;
+import com.wholeseeds.mindle.domain.region.repository.custom.SubdistrictRepositoryCustom;
 
 import jakarta.persistence.EntityManager;
