+++ src/main/java/com/wholeseeds/mindle/domain/region/service/LocationService.java
@@ -1,15 +1,15 @@
-package com.wholeseeds.mindle.domain.location.service;
+package com.wholeseeds.mindle.domain.region.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.wholeseeds.mindle.domain.complaint.dto.request.SaveComplaintRequestDto;
-import com.wholeseeds.mindle.domain.location.entity.City;
-import com.wholeseeds.mindle.domain.location.entity.Subdistrict;
-import com.wholeseeds.mindle.domain.location.exception.CityNotFoundException;
-import com.wholeseeds.mindle.domain.location.exception.SubdistrictNotFoundException;
-import com.wholeseeds.mindle.domain.location.repository.CityRepository;
-import com.wholeseeds.mindle.domain.location.repository.SubdistrictRepository;
+import com.wholeseeds.mindle.domain.region.entity.City;
+import com.wholeseeds.mindle.domain.region.entity.Subdistrict;
+import com.wholeseeds.mindle.domain.region.exception.CityNotFoundException;
+import com.wholeseeds.mindle.domain.region.exception.SubdistrictNotFoundException;
+import com.wholeseeds.mindle.domain.region.repository.CityRepository;
+import com.wholeseeds.mindle.domain.region.repository.SubdistrictRepository;
 import com.wholeseeds.mindle.domain.place.entity.Place;
 import com.wholeseeds.mindle.domain.place.exception.PlaceNotFoundException;
 import com.wholeseeds.mindle.domain.place.repository.PlaceRepository;
